Add binary chart with fun quality+quantity
Quality represented by point either at 1 or -1.
Quantity represented by size of bubble around the point.

These work well for displaying all the data we have
(based on Turbulence 3 and Bigfoot Country so far)
and they can be combined with Shmentina and Capellyana
on the same graph (yet to complete the code for this).

y is not an accurate representation of time ...yet