Scaffolding for user registration
* Create a /register page

* Check that username doesn't already exist.
Does this by trying to access Pony's list,
"if username in User" syntactic sugar doesn't exist though,
so add on-success functionality to the except block.
Nicely moves error condition to the top of the function go-style!
(I have a feeling I will question my decision later :D)

* After successful registration, redirect the user to the login page
with a message to let them know what's going on.
Use HTTP 303, which seems to exist for exactly this purpose, see
https://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html#sec10.3.4

* No database write yet.